Implemented log in and register functions on the LogInActivity.java class. Still need to properly comment code and change log in success function to lead to the new activity based on user type.

+++ app/src/main/java/com/example/cop4331projectdefinitive/User.java
@@ -8,7 +8,7 @@ import com.google.gson.JsonElement;
 
 @Entity
 public class User {
-    public User(String userName, int password, String userType) {
+    public User(String userName, double password, String userType) {
         this.userName = userName;
         this.password = password;
         this.userType = userType;
@@ -22,7 +22,7 @@ public class User {
     public String userName;
 
     @ColumnInfo(name = "pass_word")
-    public int password;
+    public double password;
 
     @ColumnInfo(name = "user_type")
     public String userType;
